99.dockter.1 (#102)
* #99 Add windows testing

* #99 Add windows testing

* #99 Add windows testing - 3

* #99 Add windows testing - 4

* #99 Savepoint

* #99 Update dependencies

* #99 Add quotes

* #99 try C.size_t

* #99 try C.size_t -

* #99 try C.size_t - 3

* #99 try C.size_t - 4

* #99 Use new GitHub action to install Windows - 1

* #99 Use new GitHub action to install Windows - 2

* #99 Use new GitHub action to install Windows - 4

* #99 Use new GitHub action to install Windows - 5

* #99 Add to Path variable

* #99 Add to Path variable

* #99 Work with environment variables

* #99 Work with environment variables - 2

* #99 Work with environment variables - 3

* #99 Work with environment variables - 4

* #99 Work with environment variables - 5

* #99 Work with environment variables - 6

* #99 Work with environment variables - 7

* #99 Work with environment variables - 8

* #99 Work with environment variables - 9

* #99 Work with environment variables - 10

* #99 Work with environment variables - 11

* #99 Work with environment variables - 12

* #99 Work with environment variables - 13

* #99 Work with environment variables - 14

* #99 Work with environment variables - 15

* #99 Prepare for versioned release

+++ g2diagnostic/g2diagnostic.go
@@ -6,7 +6,9 @@ package g2diagnostic
 /*
 #include "g2diagnostic.h"
 #cgo CFLAGS: -g -I/opt/senzing/g2/sdk/c
+#cgo windows CFLAGS: -g -I"C:/Program Files/Senzing/g2/sdk/c"
 #cgo LDFLAGS: -L/opt/senzing/g2/lib -lG2
+#cgo windows LDFLAGS: -L"C:/Program Files/Senzing/g2/lib" -lG2
 */
 import "C"
 
@@ -131,7 +133,7 @@ func (client *G2diagnostic) getLastException(ctx context.Context) (string, error
 		defer func() { client.traceExit(32, result, err, time.Since(entryTime)) }()
 	}
 	stringBuffer := client.getByteArray(initialByteArraySize)
-	C.G2Diagnostic_getLastException((*C.char)(unsafe.Pointer(&stringBuffer[0])), C.ulong(len(stringBuffer)))
+	C.G2Diagnostic_getLastException((*C.char)(unsafe.Pointer(&stringBuffer[0])), C.size_t(len(stringBuffer)))
 	// if result == 0 { // "result" is length of exception message.
 	// 	err = client.getLogger().Error(4014, result, time.Since(entryTime))
 	// }
@@ -305,7 +307,7 @@ func (client *G2diagnostic) FetchNextEntityBySize(ctx context.Context, entityLis
 		defer func() { client.traceExit(10, responseResult, err, time.Since(entryTime)) }()
 	}
 	stringBuffer := client.getByteArray(initialByteArraySize)
-	result := C.G2Diagnostic_fetchNextEntityBySize_helper(C.uintptr_t(entityListBySizeHandle), (*C.char)(unsafe.Pointer(&stringBuffer[0])), C.ulong(len(stringBuffer)))
+	result := C.G2Diagnostic_fetchNextEntityBySize_helper(C.uintptr_t(entityListBySizeHandle), (*C.char)(unsafe.Pointer(&stringBuffer[0])), C.size_t(len(stringBuffer)))
 	if result < 0 {
 		err = client.newError(ctx, 4004, result, time.Since(entryTime))
 	}
