+++ g2product/g2product.go
@@ -6,7 +6,9 @@ package g2product
 /*
 #include "g2product.h"
 #cgo CFLAGS: -g -I/opt/senzing/g2/sdk/c
+#cgo windows CFLAGS: -g -I"C:/Program Files/Senzing/g2/sdk/c"
 #cgo LDFLAGS: -L/opt/senzing/g2/lib -lG2
+#cgo windows LDFLAGS: -L"C:/Program Files/Senzing/g2/lib" -lG2
 */
 import "C"
 
@@ -131,7 +133,7 @@ func (client *G2product) getLastException(ctx context.Context) (string, error) {
 		defer func() { client.traceExit(6, result, err, time.Since(entryTime)) }()
 	}
 	stringBuffer := client.getByteArray(initialByteArraySize)
-	C.G2Product_getLastException((*C.char)(unsafe.Pointer(&stringBuffer[0])), C.ulong(len(stringBuffer)))
+	C.G2Product_getLastException((*C.char)(unsafe.Pointer(&stringBuffer[0])), C.size_t(len(stringBuffer)))
 	// if result == 0 { // "result" is length of exception message.
 	// 	err = client.getLogger().Error(4002, result, time.Since(entryTime))
 	// }
